backend/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-Mail==0.9.1
humanize==4.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg==3.1.14
psycopg-binary==3.1.14
psycopg-pool==3.2.0
PyJWT==2.8.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==3.0.1
